simple#shell: restore b:interpreter correctly
b: is relative to the current buffer, and simpl#repl has the side-effect
of *changing the current buffer*!

This resulted in a hard-to-diagnose bug where the shell's buffer got the
old b:interpreter, and the old buffer completely lost the variable.

The fix is rather simple: set the correct buffer's local variable with
setbufvar (b: prefix automatically added).